+++ database.go
@@ -2,10 +2,7 @@ package main
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
-	"os"
-	"strconv"
 	"time"
 
 	"github.com/ghodss/yaml"
@@ -14,18 +11,6 @@ import (
 	"gorm.io/gorm/logger"
 	"gorm.io/gorm/schema"
 )
-
-type DatabaseConfig struct {
-	Host            string
-	Port            string
-	User            string
-	Password        string
-	Name            string
-	MaxIdleConns    int
-	MaxOpenConns    int
-	MaxConnLifetime time.Duration
-	Log             bool
-}
 
 func (p *Project) MarshalJSON() ([]byte, error) {
 	type Alias Project
@@ -67,13 +52,13 @@ func (b *Build) MarshalJSON() ([]byte, error) {
 	})
 }
 
-func openDatabase(config DatabaseConfig) (*gorm.DB, error) {
+func openDatabase(config DBConfig) (*gorm.DB, error) {
 	const retryDelay = 2 * time.Second
 	const maxAttempts = 3
-	psqlInfo := fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=postgres sslmode=disable",
+	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=postgres sslmode=disable",
 		config.Host,
 		config.Port,
-		config.User,
+		config.Username,
 		config.Password)
 
 	var db *gorm.DB
@@ -97,10 +82,10 @@ func openDatabase(config DatabaseConfig) (*gorm.DB, error) {
 
 	db.Exec(fmt.Sprintf("CREATE DATABASE %s;", config.Name))
 
-	psqlInfo = fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=disable",
+	psqlInfo = fmt.Sprintf("host=%s port=%d user=%s password=%s dbname=%s sslmode=disable",
 		config.Host,
 		config.Port,
-		config.User,
+		config.Username,
 		config.Password,
 		config.Name)
 
@@ -132,74 +117,9 @@ func openDatabase(config DatabaseConfig) (*gorm.DB, error) {
 	return db, err
 }
 
-func getLogger(config DatabaseConfig) logger.Interface {
+func getLogger(config DBConfig) logger.Interface {
 	if config.Log {
 		return logger.Default.LogMode(logger.Info)
 	}
 	return logger.Default.LogMode(logger.Silent)
-}
-
-func getDatabaseConfigFromEnvironment() (DatabaseConfig, error) {
-	conf := DatabaseConfig{}
-	var ok bool
-
-	conf.Host, ok = os.LookupEnv("DBHOST")
-	if !ok {
-		return conf, errors.New("DBHOST environment variable required but not set")
-	}
-
-	conf.Port, ok = os.LookupEnv("DBPORT")
-	if !ok {
-		return conf, errors.New("DBPORT environment variable required but not set")
-	}
-
-	conf.User, ok = os.LookupEnv("DBUSER")
-	if !ok {
-		return conf, errors.New("DBUSER environment variable required but not set")
-	}
-
-	conf.Password, ok = os.LookupEnv("DBPASS")
-	if !ok {
-		return conf, errors.New("DBPASS environment variable required but not set")
-	}
-
-	conf.Name, ok = os.LookupEnv("DBNAME")
-	if !ok {
-		return conf, errors.New("DBNAME environment variable required but not set")
-	}
-
-	var err error
-	maxIdle, ok := os.LookupEnv("DBMAXIDLECONNS")
-	if ok {
-		conf.MaxIdleConns, err = strconv.Atoi(maxIdle)
-		if err != nil {
-			return conf, err
-		}
-	} else {
-		conf.MaxIdleConns = 2 // Current default in sql package according to docs https://golang.org/pkg/database/sql/#DB.SetMaxIdleConns
-	}
-
-	maxOpen, ok := os.LookupEnv("DBMAXOPENCONNS")
-	if ok {
-		conf.MaxOpenConns, err = strconv.Atoi(maxOpen)
-		if err != nil {
-			return conf, err
-		}
-	} else {
-		conf.MaxOpenConns = 0 // Current default in sql package according to docs https://golang.org/pkg/database/sql/#DB.SetMaxOpenConns
-	}
-
-	maxLifetime, ok := os.LookupEnv("DBMAXCONNLIFETIME")
-	if ok {
-		conf.MaxConnLifetime, err = time.ParseDuration(maxLifetime)
-		if err != nil {
-			return conf, err
-		}
-	} else {
-		conf.MaxConnLifetime = 20 * time.Minute
-	}
-
-	conf.Log, _ = strconv.ParseBool(os.Getenv("DBLOG"))
-
-	return conf, nil
 }