+++ internal/testutil/testutil.go
@@ -1,0 +1,13 @@
+package testutil
+
+import (
+	"os"
+	"testing"
+)
+
+// Setenv sets the environment variable, and then unsets it when the test has
+// finished.
+func Setenv(t *testing.T, key, value string) {
+	os.Setenv(key, value)
+	t.Cleanup(func() { os.Unsetenv(key) })
+}